Extract compaction_helpers
